+++ src/components/ui/PriorityIcon.tsx
@@ -1,10 +1,14 @@
 import React from 'react';
 
-// mui
+// icons
 import FlagOutlinedIcon from '@mui/icons-material/FlagOutlined';
 import FlagIcon from '@mui/icons-material/Flag';
+
+// data
 import { indigo, purple, red } from '@mui/material/colors';
-import { PriorityType } from '../types';
+
+// types
+import { PriorityType } from '@/types';
 
 const flagColors = [null, red[500], purple[500], indigo[500], indigo[500]];
 
@@ -23,5 +27,3 @@ export const PriorityIcon: React.FC<Props> = ({ priority }) => {
     return <FlagIcon sx={{ display: 'block', color: flagColors[priority] }} />;
   }
 };
-
-export default PriorityIcon;
